Moved helper functions to indep. file
